pyproject.toml
[tool.poetry]
name = "practice-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "practice_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-bootstrap5 = "^23.3"
django-pygments-renderer = "^0.0.1"
django-debug-toolbar = "^4.2.0"
django-taggit = "^5.0.1"
taggit-selectize = "^2.11.0"
six = "^1.16.0"
django-bootstrap-icons = "^0.8.7"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
